+++ src/errors/HashException.js
@@ -1,8 +1,8 @@
 module.exports = class ExpectationTypeException extends Error {
-  constructor () {
+  constructor (message = 'Invalid security hash') {
     super()
     this.name = 'HashException'
     this.status = 403
-    this.message = 'Invalid security hash'
+    this.message = message
   }
 }